requirements.txt
fastapi
uvicorn
openai
PyMuPDF
numpy
fpdf
streamlit
requests
python-dotenv
duckduckgo-search
chromadb
pypdf
langchain-text-splitters